all fix00

+++ libft00/ft_pow.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_pow.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anonymous <anonymous@student.42.fr>        +#+  +:+       +#+        */
+/*   By: achernys <achernys@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/04 14:12:07 by anonymous         #+#    #+#             */
-/*   Updated: 2017/11/06 23:15:37 by achernys         ###   ########.fr       */
+/*   Updated: 2017/11/11 07:57:16 by achernys         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long	ft_pow(long num, int pow)
+long	ft_pow(int num, int pow)
 {
 	if (pow == 0)
 		return (1);
